assignment4/requirements.txt
gunicorn>=20.1.0
jwt>=1.3.1
Flask-JSON>=0.3.4
psycopg2-binary>=2.9.1
bcrypt>=3.2.0
boto3>=1.19.12
